cv_addition/requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
streamlit==1.28.0
opencv-python==4.8.1.78
librosa==0.10.1
numpy==1.24.3
pandas==2.1.1
pillow==10.0.1
transformers==4.35.0
soundfile==0.12.1
matplotlib==3.8.0
seaborn==0.13.0
pymongo==4.5.0
python-dotenv==1.0.0